clear contacts
